+++ packages/gatsby-transformer-remark/src/extend-node-type.js
@@ -62,6 +62,14 @@ const withPathPrefix = (url, pathPrefix) =>
 // field's resolver is safe to run in another thread
 const workerPlugin = `gatsby-transformer-remark`
 
+// TODO: remove this check with next major release
+const safeGetCache = ({ getCache, cache }) => id => {
+  if (!getCache) {
+    return cache
+  }
+  return getCache(id)
+}
+
 /**
  * Map that keeps track of generation of AST to not generate it multiple
  * times in parallel.
@@ -71,7 +79,16 @@ const workerPlugin = `gatsby-transformer-remark`
 const ASTPromiseMap = new Map()
 
 module.exports = (
-  { type, pathPrefix, getNode, getNodesByType, cache, reporter },
+  {
+    type,
+    pathPrefix,
+    getNode,
+    getNodesByType,
+    cache,
+    getCache: possibleGetCache,
+    reporter,
+    ...rest
+  },
   pluginOptions
 ) => {
   if (type.name !== `MarkdownRemark`) {
@@ -79,6 +96,8 @@ module.exports = (
   }
   pluginsCacheStr = pluginOptions.plugins.map(p => p.name).join(``)
   pathPrefixCacheStr = pathPrefix || ``
+
+  const getCache = safeGetCache({ cache, getCache: possibleGetCache })
 
   return new Promise((resolve, reject) => {
     // Setup Remark.
@@ -154,7 +173,9 @@ module.exports = (
               files: fileNodes,
               getNode,
               reporter,
-              cache,
+              cache: getCache(plugin.name),
+              getCache,
+              ...rest,
             },
             plugin.pluginOptions
           )
@@ -223,7 +244,9 @@ module.exports = (
               files: fileNodes,
               pathPrefix,
               reporter,
-              cache,
+              cache: getCache(plugin.name),
+              getCache,
+              ...rest,
             },
             plugin.pluginOptions
           )